Move `PtrEq` into `Operators` lin pass
