+++ clippy_lints/src/lib.register_lints.rs
@@ -436,6 +436,7 @@ store.register_lints(&[
     operators::MODULO_ONE,
     operators::NEEDLESS_BITWISE_BOOL,
     operators::OP_REF,
+    operators::PTR_EQ,
     operators::VERBOSE_BIT_MASK,
     option_env_unwrap::OPTION_ENV_UNWRAP,
     option_if_let_else::OPTION_IF_LET_ELSE,
@@ -455,7 +456,6 @@ store.register_lints(&[
     ptr::INVALID_NULL_PTR_USAGE,
     ptr::MUT_FROM_REF,
     ptr::PTR_ARG,
-    ptr_eq::PTR_EQ,
     ptr_offset_with_cast::PTR_OFFSET_WITH_CAST,
     pub_use::PUB_USE,
     question_mark::QUESTION_MARK,
